% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSurface.R
\name{genSurface}
\alias{genSurface}
\title{Generate response surface for 2D interaction}
\usage{
genSurface(
  x,
  int,
  y = NULL,
  fit = NULL,
  read.forest = NULL,
  rectangles = NULL,
  wt.node = "size",
  varnames = colnames(x),
  nbin = 100,
  bins = NULL,
  binFun = NULL,
  yscale = 1,
  yFun = function(x) return(x),
  filter.rules = NULL
)
}
\arguments{
\item{x}{numeric feature matrix, with replicate features grouped}

\item{int}{signed interaction to plot. Formatted as 'X1+_X2+_X3-_...'}

\item{y}{response vector to be visualzed. If NULL, genSurface will use random
forest predictions associated with each decision rule}

\item{fit}{a fitted random forest, from packages randomForest or ranger.}

\item{read.forest}{output of readForest.}

\item{rectangles}{a list of hyperrectangles corresponding to leaf nodes in an
RF, as retuned by forestHR. If both rectangles and read.forest are supplied,
read.forest will be ignored.}

\item{wt.node}{indicator for how nodes are to be weighted in response
surfaces. One of `size` - weighting proportional to leaf node size or
`none` - indicating uniform weighting.}

\item{varnames}{character vector indicating feature names. If NULL,
colnames(x) are used as feature names.}

\item{nbin:}{number of bins to plot surface map over}

\item{bins:}{user generated grid to plot over. If supplied, nbin is 
ignored}

\item{filter.rules:}{a list of filtering functions to be applied to rf
decision paths. If NULL, default rules will filter to a random sample of
10% of leaf nodes with at least 5 observations.}
}
\description{
Response surface give E(Y|X) over regions learned by an iRF.
}
