% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotInt}
\alias{plotInt}
\title{Plot interaction}
\usage{
plotInt(
  x,
  int,
  y = NULL,
  fit = NULL,
  read.forest = NULL,
  varnames = colnames(x),
  col.pal = magma,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  slab = NULL,
  z.range = NULL,
  nbin = 50,
  binFun = NULL,
  yscale = 1,
  yFun = function(x) return(x),
  filter.rules = NULL,
  filterX = NULL,
  wt.node = "size",
  type = "plotly",
  main = NULL
)
}
\arguments{
\item{x}{numeric feature matrix, with replicate features grouped}

\item{int}{signed interaction to plot. If numeric, int is assumed to
correspond to column indices to be plotted for interaction. If character,
assumed to be formatted as 'X1+_X2+_X3-_...'}

\item{y}{response vector.}

\item{fit}{a fitted random forest, from packages randomForest or ranger.}

\item{read.forest}{output of readForest}

\item{varnames}{character vector indicating feature names. If NULL,
colnames(x) are used as feature names.}

\item{col.pal}{color palette for response surfaces. A function that takes an
integer input and returns colors to be use in the palette.}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{zlab}{z-axis label}

\item{slab}{for order 3 and 4 interactions, label for split plots}

\item{z.range}{z-axiz range}

\item{wt.node}{indicator for how nodes are to be weighted in response
surfaces. One of `size` - weighting proportional to leaf node size or
`none` - indicating uniform weighting.}

\item{type}{one of `rgl` - 3d response surface or ggplot - 2d response
surface}

\item{main}{plot title for response surfaces}

\item{nbin:}{number of bins to plot surface map over}

\item{min.surface}{minimum number of observations required to generate a
response surface.}

\item{filter.rules:}{a list of filtering functions to be applied to rf
decision paths. If NULL, default rules will filter to a random sample of
10% of leaf nodes with at least 5 observations.}

\item{filter_x}{a filtering function to be applied to data matrix. Takes as
arguments x (data matrix), int (numeric vector of interaciton ids), and
thresholds (numeric vector of rf thresholds, columns corresponding to
features in int), returns indices of x to be kept.}
}
\description{
Visualization of response surface plots.
}
